applications/data_processing_api/data-processing-api-requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
flasgger==0.9.7b2
python-dotenv==1.0.1
elasticsearch==7.15.0
pandas==1.3.5
GeoAlchemy2==0.15.2
scikit-learn==1.3.0
shapely==2.0.6
geopy==2.4.1
matplotlib==3.7.5
folium==0.17.0
